requirements.txt
aws-cdk-lib==2.205.0
constructs>=10.0.0,<11.0.0
boto3
aws-cdk.aws-s3
aws-cdk.aws-lambda
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-iam
